requirements.txt
streamlit>=1.28.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
plotly>=5.15.0
dnspython>=2.4.0
python-whois>=0.8.0
